Add files via upload
Write a simple grep like simple utility in the C programming language. Approach the problem by following these steps: 

a)	Your task is to write a simple grep like utility in the C programming language. You can name it “mygrep”. grep is basically used for searching. For example, “mygrep foo myfile” command returns all the lines that contain a string matching the expression "foo" in the file. Your code will be checked by running such a simple command.

It is easy to get overwhelmed with the complexity of grep. Note that we are looking for a simple version that can search for a specific string in a single file. You may consult grep man page if you are looking to get more info about it