Add files via upload
/*
Write a program which ask the user to enter the number of processes and the resources required by each process.
This program also asks for maximum resources required by each process and the allocated and available resources.
Now it must make sure that it prevents deadlock to occur.  Your program should tell which rule/ condition has been violated if the deadlock occur. 
*/